array: Provide sane defaults for GObject-style construction
The defaults is such that we at least have an array of 'gvoid' with an
initial dimension of '0'. It's usually not possible to create such an
array via the default constructor, but some bindings might want to
define the actual shape later.

Add some pre-conditions for ensuring the shape and dimension are
correct.

Fix 'reshape' tests because we need to ensure we can reshape an initial
array.