array: Expose 'origin' as 'size_t', but back it with a pointer
From a public perspective, passing a pointer does not always make sense.
In the case of a 'null' data, there would be no way to know in advance
where it would be allocated.

Instead pass a 'size_t' and compute an internal '_baseptr'.

In 'Iterator', just recompute '_baseptr' from the array's data and
origin.