Introduce 'Gradient' for retreiving partial derivatives
Each routine will have, as a convention, its gradient defined as a set
of routines computing its partial derivatives named as
'<routine>_gradient_<out>_<in>'.

The identifier and symbol can be overwritten by specifying
'vast.gradient_<out>_<in>' and 'vast.gradient_<out>_<in>_cname' GIR
attributes.

Introduce other math routines to compute necessary steps for derivating
powers (x ^ a) and exponentiation (a ^ x) where 'x' is a variable and
'a' a constant.